+++ src/main/java/chatbot/ChatbotException.java
@@ -1,3 +1,5 @@
+package chatbot;
+
 public class ChatbotException extends Exception{
     public ChatbotException(String errorMessage) {
         super(errorMessage);
