+++ src/main/java/chatbot/command/AddEventCommand.java
@@ -1,3 +1,9 @@
+package chatbot.command;
+
+import chatbot.Ui;
+import chatbot.ChatbotException;
+import chatbot.task.TaskManager;
+
 public class AddEventCommand extends Command{
     private String description;
     private String start;
