+++ src/main/java/chatbot/command/MarkCommand.java
@@ -1,13 +1,19 @@
-public class UnmarkCommand extends Command {
+package chatbot.command;
+
+import chatbot.Ui;
+import chatbot.ChatbotException;
+import chatbot.task.TaskManager;
+
+public class MarkCommand extends Command {
     private int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskManager taskManager, Ui ui) throws ChatbotException {
-        taskManager.unMarktask(index);
+        taskManager.taskDone(index);
     }
 
     @Override
