+++ src/main/java/chatbot/command/UnknownCommand.java
@@ -1,3 +1,9 @@
+package chatbot.command;
+
+import chatbot.Ui;
+import chatbot.ChatbotException;
+import chatbot.task.TaskManager;
+
 public class UnknownCommand extends Command{
     private String userInput;
