+++ src/main/java/chatbot/storage/Storage.java
@@ -1,6 +1,7 @@
+package chatbot.storage;
+
 import java.io.*;
-import java.util.ArrayList;
-import java.util.List;
+import chatbot.task.*;
 
 public class Storage {
     private final String FILEPATH;
